validators into a single file and updated
Moved EmailValidator, NameValidator, and PasswordValidator into a unified validator.dart file under a new directory. Removed the old individual validator files and updated all relevant imports in login_page.dart and signup_page.dart to use the new consolidated validator file for improved maintainability.
