--- lib/features/auth/valid/email_validator.dart
@@ -1,11 +1,0 @@
-class EmailValidator {
-  static String? validate(String? value){
-              if (value == null || value.isEmpty) {
-                return 'Email is required';
-              }
-              if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value)) {
-                return 'Enter a valid email';
-              }
-              return null;            
-  }
-}
