--- lib/features/auth/valid/name_validator.dart
@@ -1,14 +1,0 @@
-class NameValidator {
-  static String? validate(String? value) {
-    if (value == null || value.trim().isEmpty) {
-      return "Name is required";
-    }
-    if (value.trim().length < 3) {
-      return "Name must be at least 3 characters";
-    }
-    if (!RegExp(r'^[a-zA-Z\s]+$').hasMatch(value.trim())) {
-      return "Name must contain only letters";
-    }
-    return null;
-  }
-}
