+++ lib/features/auth/validator/validator.dart
@@ -1,3 +1,14 @@
+class EmailValidator {
+  static String? validate(String? value){
+              if (value == null || value.isEmpty) {
+                return 'Email is required';
+              }
+              if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value)) {
+                return 'Enter a valid email';
+              }
+              return null;            
+  }
+}
 class PasswordValidator {
   static String? validate(String? value) {
     if (value == null || value.isEmpty) {
@@ -26,6 +37,20 @@ class PasswordValidator {
     
     if (!regex.hasMatch(value)) {
       return 'Password must have uppercase, number, and special symbol (@\$!%*?&)';
+    }
+    return null;
+  }
+}
+class NameValidator {
+  static String? validate(String? value) {
+    if (value == null || value.trim().isEmpty) {
+      return "Name is required";
+    }
+    if (value.trim().length < 3) {
+      return "Name must be at least 3 characters";
+    }
+    if (!RegExp(r'^[a-zA-Z\s]+$').hasMatch(value.trim())) {
+      return "Name must contain only letters";
     }
     return null;
   }
